Move client HTTP enc/dec to svc/client/http/client.go
